Adding origin verification for server

+++ server/alerting/notification/src/main/java/com/clemble/casino/server/notification/controller/PlayerNotificationServiceController.java
@@ -4,6 +4,7 @@ import com.clemble.casino.WebMapping;
 import com.clemble.casino.notification.PlayerNotification;
 import com.clemble.casino.player.PlayerNotificationWebMapping;
 import com.clemble.casino.player.service.PlayerNotificationService;
+import com.clemble.casino.server.notification.ServerPlayerNotification;
 import com.clemble.casino.server.notification.repository.PlayerNotificationRepository;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
@@ -27,7 +28,6 @@ public class PlayerNotificationServiceController implements PlayerNotificationSe
         throw new UnsupportedOperationException();
     }
 
-
     @RequestMapping(method = RequestMethod.GET, value = PlayerNotificationWebMapping.MY_NOTIFICATIONS, produces = WebMapping.PRODUCES)
     @ResponseStatus(value = HttpStatus.OK)
     public PlayerNotification[] myNotifications(@CookieValue("player") String player) {
@@ -47,6 +47,12 @@ public class PlayerNotificationServiceController implements PlayerNotificationSe
     @RequestMapping(method = RequestMethod.DELETE, value = PlayerNotificationWebMapping.MY_NOTIFICATIONS_DELETE)
     @ResponseStatus(value = HttpStatus.NO_CONTENT)
     public void delete(@CookieValue("player") String player, @PathVariable("key") String key) {
+        // Step 1. Fetching and checking notifications
+        ServerPlayerNotification notification = notificationRepository.findOne(key);
+        if (!notification.getPlayer().equals(player))
+            throw new IllegalAccessError();
+        // Step 2. Removing by key
         notificationRepository.delete(key);
     }
+
 }
